fix: archivos por errores de eslint con vinculación a la api
