+++ src/contexts/GenderContext.jsx
@@ -1,13 +1,3 @@
-import { createContext, useState } from "react";
+import { createContext } from "react";
 
 export const GenderContext = createContext();
-
-export function GenderProvider({ children }) {
-  const [selectedGender, setSelectedGender] = useState("");
-
-  return (
-    <GenderContext.Provider value={{ selectedGender, setSelectedGender }}>
-      {children}
-    </GenderContext.Provider>
-  );
-}
