+++ vite.config.js
@@ -1,7 +1,11 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
-import path from "path";
+import { fileURLToPath } from "url";
+import { dirname, resolve } from "path";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -11,11 +15,11 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      "@assets": path.resolve(__dirname, "./src/assets"),
-      "@icons": path.resolve(__dirname, "./src/assets/icons"),
-      "@components": path.resolve(__dirname, "./src/components"),
-      "@images": path.resolve(__dirname, "./src/images"),
-      "@utils": path.resolve(__dirname, "./src/utils"),
+      "@assets": resolve(__dirname, "./src/assets"),
+      "@icons": resolve(__dirname, "./src/assets/icons"),
+      "@components": resolve(__dirname, "./src/components"),
+      "@images": resolve(__dirname, "./src/images"),
+      "@utils": resolve(__dirname, "./src/utils"),
     },
   },
 });
